third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/masonry-grid-template-columns-computed-withcontent-expected.txt

This is a testharness.js-based test.
[FAIL] Property grid-template-columns value 'none'
  assert_equals: expected "300px" but got "none"
[FAIL] Property grid-template-columns value '20%'
  assert_equals: expected "60px" but got "20%"
[FAIL] Property grid-template-columns value 'calc(30% + 40px)'
  assert_equals: expected "130px" but got "calc(30% + 40px)"
[FAIL] Property grid-template-columns value '5fr'
  assert_equals: expected "300px" but got "5fr"
[FAIL] Property grid-template-columns value 'min-content'
  assert_equals: expected "30px" but got "min-content"
[FAIL] Property grid-template-columns value 'max-content'
  assert_equals: expected "30px" but got "max-content"
[FAIL] Property grid-template-columns value 'auto'
  assert_equals: expected "300px" but got "auto"
[FAIL] Property grid-template-columns value 'minmax(10px, auto)'
  assert_equals: expected "300px" but got "minmax(10px, auto)"
[FAIL] Property grid-template-columns value 'minmax(20%, max-content)'
  assert_equals: expected "60px" but got "minmax(20%, max-content)"
[FAIL] Property grid-template-columns value 'minmax(min-content, calc(-0.5em + 10px))'
  assert_equals: expected "30px" but got "minmax(min-content, 0px)"
[FAIL] Property grid-template-columns value 'minmax(auto, 0)'
  assert_equals: expected "30px" but got "minmax(auto, 0px)"
[FAIL] Property grid-template-columns value 'fit-content(70px)'
  assert_equals: expected "30px" but got "fit-content(70px)"
[FAIL] Property grid-template-columns value 'fit-content(20%)'
  assert_equals: expected "30px" but got "fit-content(20%)"
[FAIL] Property grid-template-columns value 'fit-content(calc(-0.5em + 10px))'
  assert_equals: expected "30px" but got "fit-content(0px)"
[FAIL] Property grid-template-columns value 'repeat(1, 10px)'
  assert_equals: expected "10px" but got "repeat(1, 10px)"
[FAIL] Property grid-template-columns value 'repeat(1, [one two] 20%)'
  assert_equals: expected "[one two] 60px" but got "repeat(1, [one two] 20%)"
[FAIL] Property grid-template-columns value 'repeat(2, minmax(10px, auto))'
  assert_equals: expected "150px 150px" but got "repeat(2, minmax(10px, auto))"
[FAIL] Property grid-template-columns value 'repeat(2, fit-content(20%) [three four] 30px 40px [five six])'
  assert_equals: expected "30px [three four] 30px 40px [five six] 30px [three four] 30px 40px [five six]" but got "repeat(2, fit-content(20%) [three four] 30px 40px [five six])"
[FAIL] Property grid-template-columns value 'min-content repeat(5, minmax(10px, auto))'
  assert_equals: expected "30px 54px 54px 54px 54px 54px" but got "min-content repeat(5, minmax(10px, auto))"
[FAIL] Property grid-template-columns value '[] 150px [] 1fr []'
  assert_equals: expected "150px 150px" but got "150px 1fr"
[FAIL] Property grid-template-columns value 'repeat(auto-fill, 200px)'
  assert_equals: expected "200px" but got "repeat(auto-fill, 200px)"
[FAIL] Property grid-template-columns value 'repeat(auto-fit, [one] 20%)'
  assert_equals: expected "[one] 60px [one] 0px [one] 0px [one] 0px [one] 0px" but got "repeat(auto-fit, [one] 20%)"
[FAIL] Property grid-template-columns value 'repeat(auto-fill, minmax(100px, 5fr) [two])'
  assert_equals: expected "100px [two] 100px [two] 100px [two]" but got "repeat(auto-fill, minmax(100px, 5fr))"
[FAIL] Property grid-template-columns value 'repeat(auto-fit, [three] minmax(max-content, 6em) [four])'
  assert_equals: expected "[three] 240px [four]" but got "repeat(auto-fit, [three] minmax(max-content, 240px) [four])"
[FAIL] Property grid-template-columns value '[one] repeat(2, minmax(50px, auto)) [two] 30px [three] repeat(auto-fill, 10px) 40px [four five] repeat(2, minmax(200px, auto)) [six]'
  assert_equals: expected "[one] 50px 50px [two] 30px [three] 10px 40px [four five] 200px 200px [six]" but got "[one] repeat(2, minmax(50px, auto)) [two] 30px [three] repeat(auto-fill, 10px) 40px [four five] repeat(2, minmax(200px, auto)) [six]"
Harness: the test ran to completion.
